Reworked showFriends(), showFriendRequests(), and showProfilePosts() into ShowProfileAction.
Added logic to handle viewing the logged-in user or the current profile user in session. Added a new action to reset the session user and hence display the logged-in user's profile.
Added an About page.
Added second menu (profileMenu) to header.jsp. This links to selected user's Profile (with stream of posts), About page, list of Friends, and Friend Requests (this option only shows in case user is viewing their own profile).